Perform clustering on all tweets (2.16) (#41)
I wasn't sure whether I should be using `fit` or `fit_predict`. I think
`fit` generates a model that `predict` can then be used on to assign
each data point to a cluster and `fit_predict` does this in one step
and returns the labels. We weren't using the result of that function
call but rather we're using attributes on the `kmeans` object so we
just need to generate the model to get the result we need so I opted
for using `fit`.

Closes #32.